chatbot_project/requirements.txt
# Build tools
setuptools>=80.0.0
wheel>=0.45.0

# Core dependencies
Django>=5.2,<6
openai>=0.27.0
requests>=2.28.2
scikit-learn>=1.3.0
joblib>=1.2.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
gunicorn
whitenoise
